:art: Created a dedicated TemplateBuilder class and filled in some of the other classes. Attempting to fix incorrect gitignore structure
