+++ .gitignore
@@ -9,5 +9,6 @@
 /.idea
 /out
 target
+
 dependency-reduced-pom.xml
 OpenTerminalProject.iml
