+++ src/main/java/org/ares/openterminal/foundation/impl/command/CreateSimpleCommand.java
@@ -5,6 +5,7 @@ import org.apache.velocity.VelocityContext;
 import org.apache.velocity.app.VelocityEngine;
 import org.ares.openterminal.Buildable;
 import org.ares.openterminal.util.StringUtil;
+import org.ares.openterminal.util.TemplateBuilder;
 import org.ares.openterminal.util.VelocityBuilder;
 import org.ares.openterminal.util.YamlHandler;
 import picocli.CommandLine.Command;
@@ -12,7 +13,7 @@ import picocli.CommandLine.Parameters;
 import java.io.*;
 
 @Command(name = "make:command")
-public class CreateSimpleCommand implements Runnable, Buildable {
+public class CreateSimpleCommand implements Runnable {
 
     @Parameters()
     private String name;
@@ -23,7 +24,6 @@ public class CreateSimpleCommand implements Runnable, Buildable {
 
     final YamlHandler yamlHandler = new YamlHandler();
 
-    final String projectPath = yamlHandler.getProjectPath();
     final String packageName = yamlHandler.getPackageName(PROPERTY_KEY);
 
     public VelocityContext buildContext() {
@@ -36,49 +36,13 @@ public class CreateSimpleCommand implements Runnable, Buildable {
         return context;
     }
 
-    // TODO: Extract to class??
-    public Writer createFileWriter() {
-        try {
-            return new FileWriter(yamlHandler.getTargetLocation(projectPath, yamlHandler.getKeyValue(PROPERTY_KEY), name) + ".java");
-        } catch (IOException ioException) {
-
-            System.out.println(ioException.getMessage());
-        }
-
-        return null;
-    }
-
-    public void flushFileWriter(Writer writer) {
-        try {
-            writer.flush();
-            writer.close();
-        } catch (IOException e) {
-            e.printStackTrace();
-        }
-    }
-
-    public void createTemplate(Writer writer) {
-        VelocityEngine engine = new VelocityBuilder().createVelocityEngineFoundation();
-        Template template = engine.getTemplate(TEMPLATE);
-
-        engine.mergeTemplate(template.getName(), "UTF-8", buildContext(), writer);
-    }
-
-    public void writeToTemplate() {
-        final Writer writer = createFileWriter();
-
-        createTemplate(writer);
-        flushFileWriter(writer);
-    }
-
-    public void buildTemplate() {
-        buildContext();
-        writeToTemplate();
-    }
-
-
     @Override
     public void run() {
-        buildTemplate();
+
+        TemplateBuilder templateBuilder = new TemplateBuilder();
+        Writer writer = templateBuilder.createFileWriter(PROPERTY_KEY, name);
+        templateBuilder.createTemplate(writer, TEMPLATE, buildContext());
+        templateBuilder.flushFileWriter(writer);
+
     }
 }
